+++ src/transform_fns/identity.rs
@@ -1,7 +1,7 @@
 use crate::transform_fns::TransformFn;
 
 /// No hashing. Useful for debugging and testing.
-pub static NO_HASHING: TransformFn<String> = identity;
+pub const NO_HASHING: TransformFn<String> = identity;
 
 fn identity(input: &str) -> String {
     String::from(input)