+++ src/transform_fns/md5.rs
@@ -4,7 +4,7 @@ use md5::{Digest, Md5};
 use crate::transform_fns::TransformFn;
 
 /// MD5-Hashing
-pub static MD5_HASHING: TransformFn<Md5Hash> = md5;
+pub const MD5_HASHING: TransformFn<Md5Hash> = md5;
 
 pub type Md5Hash = Output<Md5>;