+++ src/transform_fns/sha1.rs
@@ -4,7 +4,7 @@ use sha1::{Digest, Sha1};
 use crate::transform_fns::TransformFn;
 
 /// Sha1-Hashing
-pub static SHA1_HASHING: TransformFn<Sha1Hash> = sha1;
+pub const SHA1_HASHING: TransformFn<Sha1Hash> = sha1;
 
 pub type Sha1Hash = Output<Sha1>;