+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - started CHANGELOG file
 - several changes to public interface
-  - all hashing algorithms now use a common trait and are not longer
+  - all hashing algorithms now use a common trait and are no longer
     copied to a string
 - multiple internal improvements
 - better code quality